requirements/prod.txt
Django==4.2
python-dotenv~=1.0.1
transliterate==1.10.2
unidecode==1.3.8
django-cleanup==8.1.0
sorl-thumbnail==12.10.0
django-ckeditor==6.7.0
pillow==10.2.0